보스 HP 연동 몬스터 총알 랜덤

+++ shikhondo/EnemyBoss.cpp
@@ -127,7 +127,7 @@ void EnemyBoss::Update()
 void EnemyBoss::Render(HDC hdc)
 {
 #ifdef _DEBUG
-	Rectangle(hdc, pos.x - (size.cx / 2), pos.y - (size.cy / 2), pos.x + (size.cx / 2), pos.y + (size.cy / 2));
+	Rectangle(hdc, (int)(pos.x - (size.cx / 2)), (int)(pos.y - (size.cy / 2)), (int)(pos.x + (size.cx / 2)), (int)(pos.y + (size.cy / 2)));
 #endif // _DEBUG
 	ImageManager::GetSingleton()->DrawAnimImage(hdc, imageinfo);
 }
